assignment-1/chapter-5/requirements.txt
numpy
pytorch
opencv
